requirements.txt
Django==1.11.4
django-filter==0.13.0
djangorestframework==3.3.3
et-xmlfile==1.0.1
gunicorn==19.7.1
httplib2==0.9.2
jdcal==1.3
Markdown==2.6.5
oauth2client==2.2.0
olefile==0.44
openpyxl==2.4.8
Pillow==4.2.1
psycopg2==2.7.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pytz==2017.2
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
wget==3.2
